// packages/core/grid/cells/cell-types.ts
export interface BaseDrawArgs {
    ctx: CanvasRenderingContext2D;
    theme: FullTheme;
    col: number;
    row: number;
    rect: Rectangle;
    highlighted: boolean;
    hoverAmount: number;
    hoverX: number | undefined;
    hoverY: number | undefined;
    cellFillColor: string;
    imageLoader: ImageWindowLoader;
    spriteManager: SpriteManager;
    hyperWrapping: boolean;
    cell: InnerGridCell;
}
// ...
export type DrawCallback<T extends InnerGridCell> = (args: DrawArgs<T>, cell: T) => void;
type PrepCallback = (args: BaseDrawArgs, lastPrep?: PrepResult) => Partial<PrepResult>;

interface BaseCellRenderer<T extends InnerGridCell> {
    // drawing
    readonly kind: T["kind"];
    readonly draw: DrawCallback<T>;
    readonly drawPrep?: PrepCallback;
    readonly needsHover?: boolean | ((cell: T) => boolean);
    readonly needsHoverPosition?: boolean;
    readonly measure?: (ctx: CanvasRenderingContext2D, cell: T, theme: FullTheme) => number;

    // editing
    readonly provideEditor?: ProvideEditorCallback<T>;

    // event callbacks
    readonly onClick?: (
        args: {
            readonly cell: T;
            readonly posX: number;
            readonly posY: number;
            readonly bounds: Rectangle;
            readonly location: Item;
            readonly theme: FullTheme;
            readonly preventDefault: () => void;
        } & BaseGridMouseEventArgs
    ) => T | undefined;

    readonly onSelect?: (
        args: {
            readonly cell: T;
            readonly posX: number;
            readonly posY: number;
            readonly bounds: Rectangle;
            readonly theme: FullTheme;
            readonly preventDefault: () => void;
        } & BaseGridMouseEventArgs
    ) => void;
    readonly onDelete?: (cell: T) => T | undefined;
}

/** @category Renderers */
export interface InternalCellRenderer<T extends InnerGridCell> extends BaseCellRenderer<T> {
    readonly useLabel?: boolean;
    readonly getAccessibilityString: (cell: T) => string;
    readonly onPaste: (
        val: string,
        cell: T,
        details: {
            // fixme this should become the only argument
            readonly rawValue: string | string[] | number | boolean | BooleanEmpty | BooleanIndeterminate;
            readonly formatted?: string | string[];
            readonly formattedString?: string; // convenience
        }
    ) => T | undefined;
}